requirements.txt
aiohttp
aiogram
openai
ffmpeg
faster-whisper
pillow
pytesseract
